+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+- Add Support for Swift Concurrency
+
+### Removed
+
+- Remove support for iOS 10, 11 and 12
+- Remove support for tvOS 10, 11 and 12
+
+### Changed
+
+- Add @MainActor attribute to Coordinator
+- Add @MainActor attribute to NativeNavigator
+- Update Example project to pass strict concurrency checks
+
 ## [1.0.2] - 2021-09-10
 
 - Add SPM compatibility